Update README.md
With April 2018 release, the location of the pre-installed picasso.js and Plugins changed. I took this opportunity to remove the require that referred to the pre-installed versions and require the one from subfolder "node_modules". This has changed, too, since Qlik no longer named the files .min.js but only .js ... make sure to update to the most recent picasso.js using npm as explained in readme.md